\name{BOLD.R}
\alias{BOLD.R}

\title{BOLD.R: A package to interface with the Barcode of Life Database through R}

\description{BOLD.R is a package which allows users to access data directly from BOLD into R via current APIs maintained by the BOLD system. Users can access private data by logging in to BOLD using BOLD.R, or they can access public data without the need to login. All data accessed using BOLD.R is stored in R with a consistent internal structure and this allows the user to employ the suite of functions provided by BOLD.R along with existing packages to perform statistical analysis on DNA sequences.

    BOLD.R is a powerful, flexible and convenient platform which provides the user with the capacity to access and analyze large amounts of private and public data on BOLD directly through R.}
\author{Nishan Mudalige, Joris D'hondt}
\keyword{package}
